+++ WaterPipe.pde
@@ -5,7 +5,7 @@ class WaterPipe extends WaterElement
   public float maxtransfer;
   public int dir;
   
-  public WaterPipe(int c1, int c2, int maxtransfer, int dir)
+  public WaterPipe(int c1, int c2, float maxtransfer, int dir)
   {
     this.c1 = c1;
     this.c2 = c2;
